@cnqpzhang 2021-01-08
/label add hotspot