@tschatzl 2021-01-08
Fwiw, the most appropriate label for this change would probably be "hotspot-gc", but "hotspot" is fine too.