@cnqpzhang 2021-01-08
> would you mind closing this PR here on openjdk/jdk and reopening a new one on openjdk/jdk16?
Understood, I will do this today.
 
> As one of the persons typically triaging new issues in the bug tracker, I would also like to ask you to not open new issues immediately. 
OK. Can I use the same tracker (8259380 for the PR to jdk16?