@cnqpzhang 2021-01-08
/label remove hotspot