@tschatzl 2021-01-08
Yes, keep everything the same, the only difference is to create a pull request for openjdk/jdk16, not openjdk/jdk.