@cnqpzhang 2021-01-08
> Yes, keep everything the same, the only difference is to create a pull request for openjdk/jdk16, not openjdk/jdk.

Done. Please review the copied https://github.com/openjdk/jdk16/pull/97, and I am going to close this as suggested. Thanks. @tschatzl